requirements.txt
fastapi
uvicorn
requests
python-dotenv

mysql-connector-python

langchain
langchain-community
langchain-core

google-generativeai

faiss-cpu


pypdf
python-docx

tiktoken

httpx